Add files via upload
This code implements an accordion feature where each button click toggles the visibility of its associated content. When a .question-btn is clicked, the event listener triggers, identifying the button's grandparent element as the container for the question and its answer. It then toggles the show-text class on this container, showing or hiding the answer. This interaction creates a user-friendly experience, allowing information to be compactly displayed and accessed on demand, ideal for FAQs or similar sections.

Youtube : https://youtu.be/qPGIs7nq4yw

LinkedIn : https://www.linkedin.com/in/naser-raoofi